0.5 update (#38)
* upgrade to DGL 0.5

* add script

+++ nets/superpixels_graph_classification/mo_net.py
@@ -78,8 +78,8 @@ class MoNet(nn.Module):
     def compute_pseudo(self, edges):
         # compute pseudo edge features for MoNet
         # to avoid zero division in case in_degree is 0, we add constant '1' in all node degrees denoting self-loop
-        srcs = 1/np.sqrt(edges.src['deg']+1)
-        dsts = 1/np.sqrt(edges.dst['deg']+1)
+        srcs = 1/torch.sqrt(edges.src['deg'].float()+1)
+        dsts = 1/torch.sqrt(edges.dst['deg'].float()+1)
         pseudo = torch.cat((srcs.unsqueeze(-1), dsts.unsqueeze(-1)), dim=1)
         return {'pseudo': pseudo}
